refactor FolderEdit
