+++ src/sidebar/FeedList/FolderEdit/folder-edit.css
@@ -1,0 +1,19 @@
+.folder-edit {
+    display: flex;
+    flex-direction: row;
+    gap: 4px;
+    padding: 1px 0;
+    height: 20px;
+
+    padding-left: 12px;
+    padding-right: 12px;
+}
+
+.folder-edit__icon {
+    flex-shrink: 0;
+    margin-top: -2px; /* align with label */
+}
+
+.folder-edit__input {
+    flex: 1;
+}
