+++ src/sidebar/FeedList/feed-list.css
@@ -1,2 +1,3 @@
 @import './Folder/folder.css';
+@import './FolderEdit/folder-edit.css';
 @import './FeedItem/feed-item.css';
