+++ src/sidebar/sidebar-styles.css
@@ -24,7 +24,7 @@
     grid-template-columns: 32px 1fr 32px 32px;
 }
 
-input {
+.text-input {
     height: 32px;
     border: 1px solid var(--control-border-color);
     border-radius: 2px;
